requirements.txt
nextcord>=2.6
PyNaCl>=1.5.0
openai
textblob
python-dotenv>=1.0.1
platformdirs>=4.2
psutil
mafic>=2.10.1
yt-dlp>=2025.1.1
Flask
icalendar
aiohttp
cachetools
ffmpeg
